updated domain classes

+++ FinalProject/eCommerce/src/main/java/miu/edu/ecommerce/domain/Role.java
@@ -21,7 +21,7 @@ public class Role implements Serializable {
     private long id;
 
     private String name;
-    public Integer getId() {
+    public long getId() {
         return id;
     }
